Refactor custom ca example
