+++ examples/customca/main.go
@@ -1,0 +1,43 @@
+package main
+
+import (
+	"crypto/tls"
+	"crypto/x509"
+	"flag"
+	"log"
+	"net/http"
+
+	"github.com/elazarl/goproxy"
+)
+
+func main() {
+	verbose := flag.Bool("v", false, "should every proxy request be logged to stdout")
+	addr := flag.String("addr", ":8080", "proxy listen address")
+	flag.Parse()
+
+	cert, err := parseCA(_caCert, _caKey)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	customCaMitm := &goproxy.ConnectAction{Action: goproxy.ConnectMitm, TLSConfig: goproxy.TLSConfigFromCA(cert)}
+	var customAlwaysMitm goproxy.FuncHttpsHandler = func(host string, ctx *goproxy.ProxyCtx) (*goproxy.ConnectAction, string) {
+		return customCaMitm, host
+	}
+
+	proxy := goproxy.NewProxyHttpServer()
+	proxy.OnRequest().HandleConnect(customAlwaysMitm)
+	proxy.Verbose = *verbose
+	log.Fatal(http.ListenAndServe(*addr, proxy))
+}
+
+func parseCA(caCert, caKey []byte) (*tls.Certificate, error) {
+	parsedCert, err := tls.X509KeyPair(caCert, caKey)
+	if err != nil {
+		return nil, err
+	}
+	if parsedCert.Leaf, err = x509.ParseCertificate(parsedCert.Certificate[0]); err != nil {
+		return nil, err
+	}
+	return &parsedCert, nil
+}
