--- examples/goproxy-customca/main.go
@@ -1,20 +1,0 @@
-package main
-
-import (
-	"flag"
-	"log"
-	"net/http"
-
-	"github.com/elazarl/goproxy"
-)
-
-func main() {
-	verbose := flag.Bool("v", false, "should every proxy request be logged to stdout")
-	addr := flag.String("addr", ":8080", "proxy listen address")
-	flag.Parse()
-	setCA(caCert, caKey)
-	proxy := goproxy.NewProxyHttpServer()
-	proxy.OnRequest().HandleConnect(goproxy.AlwaysMitm)
-	proxy.Verbose = *verbose
-	log.Fatal(http.ListenAndServe(*addr, proxy))
-}
